chore(web-search): rename requester test helper
