+++ src/backend/services/webSearch/providers/api/__tests__/BochaProvider.test.ts
@@ -3,7 +3,7 @@ import type { WebSearchProvider, WebSearchExecutionConfig } from '@/shared/data/
 import { ApiKeyRotationState } from '../../../utils/provider';
 import bochaResponse from '../../__tests__/fixtures/bocha-response.json';
 import { BochaProvider } from '../BochaProvider';
-import { createMockJsonRequester } from './_webSearchProviderTestUtils';
+import { createMockJsonRequester } from './webSearchJsonRequesterMocks';
 
 const runtimeConfig: WebSearchExecutionConfig = {
   maxResults: 4,
