+++ src/backend/services/webSearch/providers/api/__tests__/QueritProvider.test.ts
@@ -7,7 +7,7 @@ import { QueritProvider } from '../QueritProvider';
 import {
   createMockJsonRequester,
   createRejectedJsonRequester,
-} from './_webSearchProviderTestUtils';
+} from './webSearchJsonRequesterMocks';
 
 const runtimeConfig: WebSearchExecutionConfig = {
   maxResults: 4,
